Added pagination and removed IO Balance plot tab
